Hooked up the populations filter to the results view.
Began refactor to come in line with Noranda's parts of the app
